+++ internal/adminapi/konnect.go
@@ -12,6 +12,7 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/kong/go-kong/kong"
 
+	"github.com/kong/kubernetes-ingress-controller/v3/internal/konnect/tracing"
 	tlsutil "github.com/kong/kubernetes-ingress-controller/v3/internal/util/tls"
 )
 
@@ -95,4 +96,17 @@ func EnsureKonnectConnection(ctx context.Context, client *kong.Client, logger lo
 	}
 
 	return nil
+}
+
+// KonnectHTTPDoer is a Doer implementation to be used with Konnect Admin API client. It decorates the HTTP client Do
+// method with extracting tracing information from the response headers and logging it for correlation with traces in
+// DataDog.
+func KonnectHTTPDoer() kong.Doer {
+	return func(ctx context.Context, client *http.Client, req *http.Request) (*http.Response, error) {
+		resp, err := tracing.DoRequest(ctx, client, req)
+		if err != nil {
+			return nil, err
+		}
+		return resp, nil
+	}
 }